Flytt HendelseRequest og Hendelse til SkattHendelseConsumer
