+++ src/main/kotlin/no/nav/melosysskattehendelser/SkatteHendelserConfig.kt
@@ -4,8 +4,8 @@ import no.nav.melosysskattehendelser.domain.PersonRepository
 import no.nav.melosysskattehendelser.melosys.consumer.VedtakHendelseConsumer
 import no.nav.melosysskattehendelser.melosys.producer.SkattehendelserProducer
 import no.nav.melosysskattehendelser.sigrun.AzureContextExchangeFilter
-import no.nav.melosysskattehendelser.sigrun.Hendelse
 import no.nav.melosysskattehendelser.sigrun.SigrunRestConsumer
+import no.nav.melosysskattehendelser.skatt.Hendelse
 import no.nav.security.token.support.client.core.oauth2.OAuth2AccessTokenService
 import no.nav.security.token.support.client.spring.ClientConfigurationProperties
 import org.springframework.beans.factory.annotation.Value
